requirements.txt.txt
psycopg2-binary
requests
shapely
schedule
folium
flask